AiModels/RecoveryAi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.54.4
httpx==0.27.2
stripe==11.2.0
email-validator==2.2.0
python-dotenv==1.0.1
firebase-admin==6.6.0
